기능구현완료

+++ community/views.py
@@ -10,19 +10,16 @@ from django.conf import settings
 # from login.models import User
 
 def writepage(request):
-    # if not request.session.get('accounts_user'):
-    #     return redirect('/accounts/login')
+    if not request.user.is_authenticated:
+        return redirect('/accounts/login')
     
     if request.method == 'POST':
         form = PostForm(request.POST)
         if form.is_valid():
-            post = form.save()
+            post = form.save(commit=False)
             post.writer = request.user
             post.save()
-            return redirect('/community/', post.id)
-        else:
-            form = PostForm()
-        return render(request, 'community/writepage.html', {'form':form})
+            return redirect('community/', post.id)
     else:
         form = PostForm()
         return render(request, 'community/writepage.html', {'form':form})
@@ -86,29 +83,32 @@ def detail(request, pk):
         'comments':comments,
     }
     if request.method == 'POST':
-        detail.delete()
-        return redirect('/community/')
+        if request.user.is_authenticted:
+            if request.user == detail.writer:
+                detail.delete()
+                return redirect('/community/')
+            return redirect('community:detail')
     else:
         return render(request, 'community/detail.html', context)
 
     
 def update(request, pk):
     detail = get_object_or_404(Board, pk=pk)
-    # # if request.user_id != detail.writer:
-    #     message.error(request, '수정 권한이 없습니다.')
-    #     return redirect('community:detail', user_id=pk)
-    if request.method == "POST":
-        form = PostUpdate(request.POST, instance=detail)
-        if form.is_valid():
-            detail.postname = form.cleaned_data['postname']
-            detail.contents = form.cleaned_data['contents']
-            detail.save()
-            return redirect('/community/detail/'+str(detail.id))
+    if request.user == detail.writer:
+        if request.method == "POST":
+            form = PostUpdate(request.POST, instance=detail)
+            if form.is_valid():
+                detail.postname = form.cleaned_data['postname']
+                detail.contents = form.cleaned_data['contents']
+                detail.save()
+                return redirect('/community/detail/'+str(detail.id))
+        else:
+            form = PostUpdate(instance=detail)
+        context = {'form':form}
+        return render(request, 'community/update.html', {'form':form})
     else:
-        form = PostUpdate(instance=detail)
-    context = {'form':form}
-    return render(request, 'community/update.html', {'form':form})
-
+        return redirect('community:detail')
+    
 
 def comments_create(request, pk):
     if request.user.is_authenticated:
